Better solutions for day 5
My previous solutions spent most of their time constructing lists. We
can avoid this by writing a finite state machine that operates on a
doubly linked list that contains indexes into the polymer list.

The machine keeps a node N in the linked list. Modulo stopping
conditions, it checks if N and its child should be eliminated. If so, it
does that and backs up one element; if not it advances one element.

This lets us process the whole polymer list in more or less a single
pass, and avoid all of the previous list allocations.

On my machine, this results in a 10x speedup. We now do part A in about
100ms, and part B in around 2.6 seconds.