Day 5 GO FAST
We don't need the doubly linked list from before; we just need two
queues of processed and not processed items.

If the next thing in the unprocessed queue pairs with the last thing in
the processed queue, we drop both. Otherwise we add to the processed
queue.

With this, we do part A in 50 ms and part B in 550 ms.